backend/requirements-api.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.0

sqlalchemy==2.0.35
psycopg[binary]>=3.2.0
alembic==1.13.2
pydantic-settings==2.5.2
python-dotenv>=1.0.0

celery==5.4.0
redis==5.0.8

numpy==2.2.2
pandas==2.2.3
scikit-learn==1.6.1
joblib==1.4.2
xgboost>=2.1.0
shap>=0.46.0
rapidfuzz>=3.9.0

httpx==0.28.1
